LayoutTests/fast/mediastream/delayed-permission-denied-expected.txt

Tests that no callbacks are invoked until permission is denied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.name is "NotAllowedError"
PASS successfullyParsed is true

TEST COMPLETE
